완성

+++ 소스코드/rsync.c
@@ -284,7 +284,10 @@ int sync_file(int argc, char *argv[], char *src, const char *dest, int toption) 
 	}
 
 	// 같은 파일이 이미 존재
-	if (is_same_file(src, dest))
+	if (is_same_file(src, dest)) {
+		return 0;
+	}
+
 	fname = strrchr(src, '/');
 	*fname = 0;
 	strcpy(path, src);
@@ -315,6 +318,11 @@ int sync_file(int argc, char *argv[], char *src, const char *dest, int toption) 
 		sprintf(buf, "tar -xf %s -C %s", backup_filepath, dest);
 		system(buf);
 
+		if (stat(backup_filepath, &statbuf) < 0) {
+			fprintf(stderr, "stat error for %s\n", backup_filepath);
+			exit(1);
+		}
+
 		// tar 파일 삭제
 		unlink(backup_filepath);
 
@@ -326,6 +334,10 @@ int sync_file(int argc, char *argv[], char *src, const char *dest, int toption) 
 		log_rsync(argc, argv, buf);
 	}
 	else {
+		if (stat(src, &statbuf) < 0) {
+			fprintf(stderr, "stat error for %s\n", src);
+			exit(1);
+		}
 		sprintf(backup_filepath, "%sXXXXXX", src);
 
 		// 임시 파일, SIGINT 발생 시 되돌리기 백업용
@@ -661,10 +673,6 @@ void sync_dir(int argc, char *argv[], const char *src, const char *dest, int rop
 		}
 	}
 
-	// 같은 파일이 이미 존재
-	if (is_same_file(src, dest))
-		return;
-
 	// src 디렉토리 읽기
 	if ((count = scandir(src, &dirp, NULL, NULL)) < 0) {
 		fprintf(stderr, "scandir error for %s\n", src);
@@ -718,7 +726,7 @@ void sync_dir(int argc, char *argv[], const char *src, const char *dest, int rop
 		}
 
 		// src가 빈 디렉토리인 경우
-		if (count == 2) {
+		if (count == 2 && !is_same_file(src, dest)) {
 			n = malloc(sizeof(node));
 			strcpy(n->fname, get_path(src, depth - 1));
 			n->stat.st_size = 0;
@@ -1054,7 +1062,6 @@ void sync_dir(int argc, char *argv[], const char *src, const char *dest, int rop
 			// 삭제 파일 리스트
 			tmp = glob_delete_list.next;
 
-			// 동기화 해야하는 파일들 다 tar로 묶음
 			while (tmp != NULL) {
 				cp += sprintf(cp, "\t%s delete\n", strchr(tmp->fname, '/') + 1);
 
@@ -1062,6 +1069,7 @@ void sync_dir(int argc, char *argv[], const char *src, const char *dest, int rop
 				tmp = tmp->next;
 				remove_node(prev);
 			}
+
 
 			buf[strlen(buf) - 1] = 0;
 			log_rsync(argc, argv, buf);
@@ -1096,7 +1104,7 @@ void log_rsync(int argc, char *argv[], const char *str) {
 
 	t = time(NULL);
 
-	if ((fp = fopen("ssu_rsync_log", "a")) == NULL) {
+	if ((fp = fopen("ssu_rsync_log", "a+")) == NULL) {
 		fprintf(stderr, "open error for ssu_rsync_log\n");
 		exit(1);
 	}
